Pull Table Pagination Actions into own file
